Improve layout of artwork metadata
The titles can be different lengths, which causes the cards to grow. The grid algorithm stetches all the boxes to fill the grid, but this left your eye bouncing up and down as you read the lines of information. As the name, date and medium tend to use only one line each, these seemed better lined up and separated from the multi-lines titles.

Use flexblox to grow the card content box to fill all space remaining after grid has stretched the cards. Use a mixture of padding and flex properties to align and space the content.